refactor html report
extracted html report to a separate module. removed all maven related stuff from html report.
