+++ cli/src/main/java/org/hjug/refactorfirst/cli/CliApplication.java
@@ -1,0 +1,13 @@
+package org.hjug.refactorfirst.cli;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CliApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CliApplication.class, args);
+    }
+
+}
