+++ report/pom.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.hjug.refactorfirst</groupId>
+        <artifactId>refactor-first</artifactId>
+        <version>0.4.1-SNAPSHOT</version>
+    </parent>
+
+    <groupId>org.hjug.refactorfirst.report</groupId>
+    <artifactId>report</artifactId>
+
+    <properties>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    </properties>
+    <dependencies>
+        <dependency>
+            <groupId>org.hjug.refactorfirst.graphdatagenerator</groupId>
+            <artifactId>graph-data-generator</artifactId>
+        </dependency>
+    </dependencies>
+
+</project>
