+++ report/src/test/java/org/hjug/refactorfirst/report/HtmlReportTest.java
@@ -1,13 +1,14 @@
-package org.hjug.mavenreport;
+package org.hjug.refactorfirst.report;
+
+import org.junit.jupiter.api.Test;
+
+import java.util.Locale;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.util.Locale;
-import org.junit.jupiter.api.Test;
+public class HtmlReportTest {
 
-public class RefactorFirstMavenReportTest {
-
-    private RefactorFirstMavenReport mavenReport = new RefactorFirstMavenReport();
+    private HtmlReport mavenReport = new HtmlReport();
 
     @Test
     void testGetOutputName() {
